requirements.txt
Flask
flask-cors
pandas
python-docx
pdfminer.six